chore: move prettyTime next to timer
